pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = ""
authors = ["bbaranov"]

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.29.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.29"
pydantic-settings = "^2.2.1"
werkzeug = "^3.0.2"
ruff = "^0.3.7"
isort = "^5.13.2"
pre-commit = "^3.7.0"
gunicorn = "^21.2.0"
python-jose = "^3.3.0"
sqlalchemy-utils = "^0.41.2"
pydantic = {extras = ["email"], version = "^2.7.0"}
redis = "^5.0.3"
argon2-cffi = "^23.1.0"
pycryptodomex = "^3.20.0"
python-multipart = "^0.0.9"
fastapi = "^0.111.0"
alembic = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
